+++ reckoner_dir/HistogramAnalyzer.h
@@ -4,13 +4,12 @@
 * This software is distributed under GNU GPL 3 license.
 *
 * Authors: Yun Heo, Maciej Dlugosz
-* 0.2.1
+* Version: 1.0
 *
 */
 
 #ifndef HISTOGRAMANALYZER_H
 #define HISTOGRAMANALYZER_H
-
 
 #include <kmc_api/kmer_api.h>
 #include <kmc_api/kmc_file.h>
@@ -18,13 +17,15 @@
 #include <vector>
 #include <cmath>
 
+
+
 class HistogramAnalyzer {
 private:
     bool buildHistogram(const std::string& kmcDatabase, const unsigned histogramSize, std::vector<uint64>& histogram);
 
 public:
-    uint64 getCutoffThreshold(const std::string& kmcDatabase);
-    uint64 getHistogramPeak(const std::string& kmcDatabase);
+    unsigned getCutoffThreshold(const std::string& kmcDatabase);
+    unsigned getHistogramPeak(const std::string& kmcDatabase);
 };
 
 #endif /* HISTOGRAMANALYZER_H */
